Sudoku into smaller pieces.
